move some module code into subpackages
