+++ app/src/main/java/org/example/age/app/config/AvsConfigModule.java
@@ -3,8 +3,8 @@ package org.example.age.app.config;
 import dagger.Module;
 import dagger.Provides;
 import org.example.age.module.client.AvsClientsConfig;
-import org.example.age.module.crypto.demo.AvsKeysConfig;
-import org.example.age.module.store.redis.RedisConfig;
+import org.example.age.module.crypto.demo.keys.AvsKeysConfig;
+import org.example.age.module.store.redis.client.RedisConfig;
 import org.example.age.service.AvsServiceConfig;
 
 /**
