+++ app/src/main/java/org/example/age/app/config/SiteAppConfig.java
@@ -5,8 +5,8 @@ import io.dropwizard.core.Configuration;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 import org.example.age.module.client.SiteClientsConfig;
-import org.example.age.module.crypto.demo.SiteKeysConfig;
-import org.example.age.module.store.redis.RedisConfig;
+import org.example.age.module.crypto.demo.keys.SiteKeysConfig;
+import org.example.age.module.store.redis.client.RedisConfig;
 import org.example.age.service.SiteServiceConfig;
 
 /** Configuration for the application. */
