+++ app/src/main/java/org/example/age/app/config/SiteConfigModule.java
@@ -3,8 +3,8 @@ package org.example.age.app.config;
 import dagger.Module;
 import dagger.Provides;
 import org.example.age.module.client.SiteClientsConfig;
-import org.example.age.module.crypto.demo.SiteKeysConfig;
-import org.example.age.module.store.redis.RedisConfig;
+import org.example.age.module.crypto.demo.keys.SiteKeysConfig;
+import org.example.age.module.store.redis.client.RedisConfig;
 import org.example.age.service.SiteServiceConfig;
 
 /**
