+++ module/crypto-demo/src/main/java/org/example/age/module/crypto/demo/DemoAvsVerifiedUserLocalizer.java
@@ -8,6 +8,7 @@ import java.util.concurrent.CompletionStage;
 import org.example.age.api.VerifiedUser;
 import org.example.age.api.crypto.Localization;
 import org.example.age.api.crypto.SecureId;
+import org.example.age.module.crypto.demo.keys.AvsKeysConfig;
 import org.example.age.service.module.crypto.AvsVerifiedUserLocalizer;
 
 /**
