+++ module/crypto-demo/src/main/java/org/example/age/module/crypto/demo/DemoSiteCryptoModule.java
@@ -3,6 +3,7 @@ package org.example.age.module.crypto.demo;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import dagger.Binds;
 import dagger.Module;
+import org.example.age.module.crypto.demo.keys.SiteKeysConfig;
 import org.example.age.service.module.crypto.AgeCertificateVerifier;
 import org.example.age.service.module.crypto.SiteVerifiedUserLocalizer;
