+++ module/crypto-demo/src/test/java/org/example/age/module/crypto/demo/keys/EccPublicKeyTest.java
@@ -1,14 +1,14 @@
-package org.example.age.module.crypto.demo;
+package org.example.age.module.crypto.demo.keys;
 
 import java.io.IOException;
 import org.example.age.module.crypto.demo.testing.ConfigKeyPair;
 import org.example.age.testing.JsonTesting;
 import org.junit.jupiter.api.Test;
 
-public final class EccPrivateKeyTest {
+public final class EccPublicKeyTest {
 
     @Test
     public void serializeThenDeserialize() throws IOException {
-        JsonTesting.serializeThenDeserialize(ConfigKeyPair.privateKey(), EccPrivateKey.class);
+        JsonTesting.serializeThenDeserialize(ConfigKeyPair.publicKey(), EccPublicKey.class);
     }
 }
