+++ module/crypto-demo/src/test/java/org/example/age/module/crypto/demo/keys/SiteKeysConfigTest.java
@@ -1,14 +1,14 @@
-package org.example.age.module.crypto.demo;
+package org.example.age.module.crypto.demo.keys;
 
 import java.io.IOException;
 import org.example.age.module.crypto.demo.testing.TestConfig;
 import org.example.age.testing.JsonTesting;
 import org.junit.jupiter.api.Test;
 
-public final class AvsKeysConfigTest {
+public final class SiteKeysConfigTest {
 
     @Test
     public void serializeThenDeserialize() throws IOException {
-        JsonTesting.serializeThenDeserialize(TestConfig.avs(), AvsKeysConfig.class);
+        JsonTesting.serializeThenDeserialize(TestConfig.site(), SiteKeysConfig.class);
     }
 }
